+++ test/feature/index.test.js
@@ -1,9 +1,10 @@
 //--------------------------------------------------------
-//-- Styleguide - Feature tests
+//-- Guidelines - Feature tests
 //--------------------------------------------------------
-import pkgDir from 'pkg-dir';
-import sass   from 'sass';
-import fss    from '@absolunet/fss';
+import validateCss from 'css-validator';
+import pkgDir      from 'pkg-dir';
+import sass        from 'sass';
+import fss         from '@absolunet/fss';
 
 
 const DIST = `${pkgDir.sync(__dirname)}/dist`;
@@ -19,13 +20,46 @@ const minimumValuesTest = (config) => {
 
 
 
-describe(`Validate the styleguide SCSS file`, () => {
+describe(`Validate the guidelines CSS file`, () => {
+	let css;
+	let validatorData;
+
+	test(`Ensure file exists`, () => {
+		expect(() => {
+			css = fss.readFile(`${DIST}/guidelines.css`, 'utf8');
+		}).not.toThrow();
+	});
+
+	test(`Ensure content is validable`, (done) => {
+		expect(() => {
+			validateCss({ text: css }, (error, data) => {
+				if (error) {
+					throw new Error('CSS validation failed');
+				}
+
+				validatorData = data;
+				done();
+			});
+		}).not.toThrow();
+	});
+
+	test(`Ensure content is valid`, () => {
+		expect(validatorData).toBeObject();
+		expect(validatorData.validity).toBeTrue();
+		expect(validatorData.errors).toStrictEqual([]);
+		expect(validatorData.warnings).toStrictEqual([]);
+	});
+});
+
+
+
+describe(`Validate the guidelines SCSS file`, () => {
 	let scss;
 	let css;
 
 	test(`Ensure file exists`, () => {
 		expect(() => {
-			scss = fss.readFile(`${DIST}/styleguide.scss`, 'utf8');
+			scss = fss.readFile(`${DIST}/guidelines.scss`, 'utf8');
 		}).not.toThrow();
 	});
 
@@ -36,7 +70,7 @@ describe(`Validate the styleguide SCSS file`, () => {
 		}).not.toThrow();
 	});
 
-	test(`Ensure styleguide generates no CSS`, () => {
+	test(`Ensure guidelines generates no CSS`, () => {
 		expect(css).toBeString();
 		expect(css).toBeEmpty();
 	});
@@ -44,13 +78,13 @@ describe(`Validate the styleguide SCSS file`, () => {
 
 
 
-describe(`Validate the styleguide JSON file`, () => {
+describe(`Validate the guidelines JSON file`, () => {
 	let json;
 	let config;
 
 	test(`Ensure file exists`, () => {
 		expect(() => {
-			json = fss.readFile(`${DIST}/styleguide.json`, 'utf8');
+			json = fss.readFile(`${DIST}/guidelines.json`, 'utf8');
 		}).not.toThrow();
 	});
 
@@ -60,23 +94,23 @@ describe(`Validate the styleguide JSON file`, () => {
 		}).not.toThrow();
 	});
 
-	test(`Ensure styleguide contains minimum values`, () => {
+	test(`Ensure guidelines contains minimum values`, () => {
 		minimumValuesTest(config);
 	});
 });
 
 
 
-describe(`Validate the entry point returns the styleguide`, () => {
-	let styleguide;
+describe(`Validate the entry point returns the guidelines`, () => {
+	let guidelines;
 
 	test(`Ensure it works`, () => {
 		expect(() => {
-			({ styleguide } = require(`${DIST}/node`));  // eslint-disable-line node/global-require
+			({ guidelines } = require(`${DIST}/node`));  // eslint-disable-line node/global-require
 		}).not.toThrow();
 	});
 
-	test(`Ensure styleguide contains minimum values`, () => {
-		minimumValuesTest(styleguide);
+	test(`Ensure guidelines contains minimum values`, () => {
+		minimumValuesTest(guidelines);
 	});
 });