+++ CHANGELOG.md
@@ -10,7 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- CSS variables output
+
 ### Changed
+- Valtechize the guidelines
+- Rename `styleguide` to `guidelines`
+- Rename `ressources` folder to `resources`
 - Maintenance update